Creating Read me
The file has been added by a Computer Science student in Technical Writing course. The student is tasked to revise or add Documentation or technical information on Github. Students will get full mars if the author accepts the changes in the pull request.

Please review my documentation and accept the request if necessary. I am ready to edit or add more to the documentation if you fell the revision not well written. 

Thank you for your time.